refactor: トークンの検証を行う処理の記述場所の修正
